Removing code smell and moving acspUpdateService to update-acsp folder
